Studia/BGT/project/requirements.txt
Flask==1.1.2
WTForms==2.3.1
flask-wtf==0.14.3
google-cloud-bigquery==1.25.0
google-cloud-tasks==1.5.0
google-cloud-storage==1.29.0
kaggle==1.5.6